requirement.txt
Flask==2.0.1
numpy==1.19.5
openpyxl==3.0.10
pandas==1.4.2
PyMySQL==1.0.2
python-dotenv==0.18.0
SQLAlchemy==1.4.37
